Move dataflow plan helper classes.
